Create replace_excerpt.R
A script for automatically logging into a WordPress website and modifying the excerpt of all posts. Utilizes RSelenium, httr, rvest, and stringr R libraries to accomplish this. The script will log into the website using the provided username and password, change the number of posts shown on the Posts page to 200, and then extract the post excerpt from each post, replacing the existing excerpt with the extracted text. Finally, the script will publish all the changes made to the posts.